Добавлена отрисовка списка ивентов через main.js, функция метода перенесена в отдельный файл
