+++ public/index.html
@@ -33,7 +33,6 @@
       <div class="page-body__container">
           <section class="trip-events">
             <h2 class="visually-hidden">Trip events</h2>
-            <ul class="trip-events__list"></ul>
           </section>
       </div>
     </main>
