+++ src/view/event-list-view.js
@@ -1,0 +1,4 @@
+export const createEventsListTemplate = () => (
+  `<ul class="trip-events__list">
+    </ul>`
+);
